examples/vdp-101/async/requirement.txt
opencv-python-headless==4.6.0.66
ffmpeg-python===0.2.0
google-cloud-storage==2.5.0
psycopg2==2.9.3
requests==2.26.0
tqdm==4.64.0